centos5/6: multiple slave refactor
We previously used to run multiple jenkins on centos5 and 6 to handle
gcc41/gcc44 jobs. This logic is now handled by passing labels in jenkins
(more specifically centos6-64-gcc44, centos5-32-gcc41). Doing so, we can
remove all the extra server stuff.

While at it, a minor fix to the centos6 init script for ccache as well
as simplifying server_id.

PR-URL: https://github.com/nodejs/build/pull/273
Reviewed-By: Joao Reis <reis@janeasystems.com>