build: changes to support ubuntu 14 for PPC BE
ubuntu PPC BE defaults to 32 bit compiles as opposed to
the rest of the distros (including PPC LE) which default to 64
bit.  This is the case even though the OS is 64 bit. The
net is that since we want to ship/support 64 bit binaries we
need to install the multi-lib package and only do that
for PPC BE.  Looking at the ansible vars there does not
seem to be one that can be used to identify between
BE versus LE so I first check that we are on PPC from the
vars which are available, and if we are PPC then use
the node name (which includes PPC64_be for be machines)
to figure out if it is BE or LE.

This change adds the required changes to support BE
ubuntu machines as well as the definition of the
initial ubuntu BE machine

PR-URL: https://github.com/nodejs/build/pull/382
Reviewed-By: Johan Bergstroem <johan@bergstroem.nu>