doc: update Vagrantfile in local ansible test guide
For some reason, the previous Vagrantfile example didn't work anymore.
When running `vagrant up` the root user was simply not able to do
the initial login to the server anymore, it stalled at:

```
==> default: Waiting for machine to boot. This may take a few minutes...
    default: SSH address: 127.0.0.1:2222
    default: SSH username: root
    default: SSH auth method: private key
    default: Warning: Authentication failure. Retrying...
    default: Warning: Authentication failure. Retrying...
    default: Warning: Authentication failure. Retrying...
    default: Warning: Authentication failure. Retrying...
```

Therefore the virtual machine setup procedure had to be manually aborted.

With these changes, we don't initially login with root anymore as that
turned out to be quite brittle. We now use the traditional 'vagrant' user
and copy the custom SSH key into root's .ssh-folder as the last step,
still allowing us to login as root after the virtual machine has been setup.