github-bot: add $LOGS_DIR env variable
We've had an issue with all logs written by the github-bot being
deleted whenever the bot is re-deployed. That's especially bad when
we want to debug issues happening a few days ago, which would be
impossible if we've done a re-deploy in between.

The root cause of that issue is that we clone the github-bot repo
onto the server and write logs to `./logs` of that cloned repo.
Whenever we re-deploy we run `git clean -fdx` before updating the repo,
resulting in any untracked files to be deleted.

By adding a `$LOGS_DIR` pointing to a directory *outside* the git repo,
will prevent logs being deleted cause of re-deploys.