Sort even versions with 4 digits
With the current version of the regex we are matching even non
semver version (eg. versions with 4 digits)

This solution seems to work with any number of digits on the version.

A possible alternative (we might prefer) is that to change the regex
only to match versions with 2/3 digits and then we can go back
using the python `StrictVersion` function.