Add virtualenv to ansible
In my Linux distribution the default python is python3, thus
I was then having issues with the python scripts in the repo
where the shabang is `#!/usr/bin/env python`.

First idea was to change the shabang to `... python2` but I
realized that a virtual env can also help to pin the versions of
ansible and libraries eg. Jinja2 and make everyone env consistent.