fix jenkins-status script
The script is broken and has been for a long time (since we made GitHub
auth mandatory to view stuff). This fixes it if and only if you have
node-core-utils configured to interact with Jenkins. Obviously, there is
room for improvement, but this takes it from "does not work at all" to
"ok, it works now at least if you've configured node-core-utils".

The alternative to maintaining the script is removing it. Seems like it
is/could be useful so keep it and land this, yeah?