ansible: set JOBS to 4 for AIX
Some of the builds were taking the 20 CPUS to mean run with JOBS=20, but
the machines don't have enough memory to run at that kind of build
concurrency. Throttle to 4. Some jobs were doing this already, but by
setting it in the global env, it should be inherited by all jobs.