doc: mv jenkins job configuration access docs
Information about the process and requirements for access to Jenkins was
in a standalone document. Merge the document into GOVERNANCE along with
the other sections on access to resources managed by the build-wg.

Its basically just a cut-n-paste, with some future-tenses changed to
current, as this is the current process.