doc: move jenkins jobs docs (#2150)
This looks to be an attempt to document the various Jenkins jobs the WG
maintains, and what they do, but it was tucked away where it was not
very visible. It doesn't describe most jobs, but it will be a bit more
discoverable in doc/, and might start getting updated.