ansible: fix z/OS ansible scripts (#2346)
Current versions of ansible checksum files before transferring them
to the remote host. Our ansible.cfg was forcing scp to be used for
transfer, which on z/OS results in an ASCII->EBCDIC conversion and
a mismatched checksum, causing ansible to error.

This commit sets the file transfer method for z/OS to sftp, which on
z/OS will transfer the file without conversion and fix the mismatched
checksums. The transferred files are either tagged as ASCII or
converted to EBCDIC if they are used by the system, e.g. the start
script.