ansible: fix curl: not found errors on AIX (#2537)
Recent versions of the AIX Toolbox curl package no longer create
symlinks in `/usr/bin`. Some jobs in the CI expect to be able to
run `curl` from the PATH.

Explicitly install the curl package (previously it was installed
as a dependency of git) and create symlinks into `/usr/bin`.