www: use docker:lts in build-site.sh
The nodejs.org website build uses `fibers` which is broken on Node.js 16
due to an incompatibility with the version of V8 that it contains.
Switch the Docker image used to build the website from `node:latest` to
`node:lts` to buy some time for the website to migrate off `fibers`.