ansible: add gcc-8 to Ubuntu 18.04 hosts/containers
Baseline gcc/g++ compiler version for Node.js 16 and above is 8. The
default `gcc` and `g++` packages on Ubuntu 18.04 install version 7.

Moves `gcc-6` and `g++-6` out of the common `ubuntu` packages into
`ubuntu1604` -- it doesn't seem worth installing a lower version of
GCC to the default for Ubuntu 18.04.

Update `select-compiler.sh` to select GCC 8 for Node.js 16 and later.