ansible: use `maint/maint-*` branches of ICU
Use the `maint/maint-*` branches of the ICU repository to build the
"system" ICU libraries. Node.js 12 requires patches on top of ICU 64
to pass intl tests which are present on the `maint/maint-64` branch
but not in the 64.2 release. These patches are present in the Node.js
source tree for `v12.x-staging` under `tools/icu/patches` but these
are only applied when building ICU statically linked.