ansible: create `gtar` symlink on AIX
New versions of the `tar` package from the AIX Toolbox no longer
create the `/usr/bin/gtar` symlink. Ansible's `unarchive` task
attempts to find `gtar` before falling back to using `tar` but on
AIX `tar` is not the GNU version of `tar` and the `unarchive` task
is incompatible with it.