ansible: AIX updates
Replace AIX 7.1 VMs. Add set up for new AIX 7.3 VM.

Differences for AIX 7.3 from AIX 7.1/7.2 set up:
- `gcc-c+` package in AIX Toolbox for AIX 7.3 is gcc 10. There are no
earlier versions of gcc available for AIX 7.3 (we install `gcc6-c+` on
AIX 7.2/7.1 but this is not available for 7.3 so we need to adjust the
playbook accordingly).
- `ccache` doesn't exist in AIX toolbox (we are compiling it from
source for AIX 7.2/7.1) but the toolbox does have a fork, `ccache-swig`,
which this PR tries to use for AIX 7.3.
- Use Java 17 instead of 8 (since this is a new machine).
- AIX toolbox is migrating to `dnf` instead of `yum`. The AIX 7.3
instance we've been given appears to already have `dnf` installed.