ansible: create git repo under `binary_tmp` user
Create the `binary_tmp` git repository under the home directory for
the `binary_tmp` user to avoid permissions issues encountered when
previously trying to create it under the `iojs` user.

This change will mean that for any machines that mount expanded
storage as an additional disk, the `binary_tmp` users' home directory
will need to be on the larger storage. In practice this means that
the additional disk should be mounted at `/home/` rather than
`/home/iojs/`.