ansible: dynamically exclude `jenkins-workspace` (#3816)
The `ansible/playbooks/jenkins/worker/create.yml` playbook runs some
tasks at the end to remove Node.js and check there is no runnable
`node` in the path. This is done on all `test` and `release` machines
with the exception of the `jenkins-workspace` machines, which need a
runnable `node` to run the linters.

Previously each `jenkins-workspace` machine was listed as an exclusion
to `hosts`. This PR moves the exclusion into the tasks and checks if
the host has an alias that begins `jenkins-workspace` (as all of our
`jenkins-workspace` machines are defined). This will avoid one place
that we have to remember to manually update every time we add/remove
a `jenkins-workspace` machine.